SEQUOIADBMAINSTREAM-2656 使用导出工具导出bindata类型数据，type值为128时出现错误
git-svn-id: http://192.168.20.11/sequoiadb/branches/sequoiadb_2.6@30276 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/util/rawbson2csv.c
@@ -74,7 +74,7 @@ typedef void (*UTIL_PLOG)( const CHAR *pFunc, \
                            const CHAR *pFmt, \
                            ... ) ;
 
-UTIL_PLOG _pPrintfLogFun = NULL ; 
+UTIL_PLOG _pPrintfLogFun = NULL ;
 
 #define UTIL_RAW2BSON_PRINTF_LOG( fmt, ... )\
 {\
@@ -526,7 +526,7 @@ INT32 _appendValue( CHAR delChar, bson_iterator *pIt,
                goto error ;
             }
             binType = (INT32)bson_iterator_bin_type( pIt ) ;
-            tempSize = ossSnprintf ( temp, 64, "%d", binType ) ;
+            tempSize = ossSnprintf ( temp, 64, "%d", (UINT8)binType ) ;
             rc = _appendString( delChar, TRUE, temp, tempSize, ppBuffer, pCSVSize ) ;
             if ( rc )
             {
